+++ C++ Library/PathfindingLib/PathfindingLib/PUnit.h
@@ -4,19 +4,19 @@
 class PUnit
 {
 public:
-	PUnit(blaze::StaticVector<double, 3> position);
+	PUnit(blaze::StaticVector<float, 3> position);
 	~PUnit();
-	void UpdatePosition(blaze::StaticVector<double, 3> position);
-	void UpdateHeading(blaze::StaticVector<double, 3> heading);
-	void UpdateTarget(blaze::StaticVector<double, 3> target);
+	void UpdatePosition(blaze::StaticVector<float, 3> position);
+	void UpdateHeading(blaze::StaticVector<float, 3> heading);
+	void UpdateTarget(blaze::StaticVector<float, 3> target);
 	void SetLeader(unsigned int leader_id);
 	unsigned int GetLeader();
-	blaze::StaticVector<double, 3> GetPosition();
-	blaze::StaticVector<double, 3> GetHeading();
-	blaze::StaticVector<double, 3> GetTarget();
+	blaze::StaticVector<float, 3> GetPosition();
+	blaze::StaticVector<float, 3> GetHeading();
+	blaze::StaticVector<float, 3> GetTarget();
 private:
-	blaze::StaticVector<double, 3> position;
-	blaze::StaticVector<double, 3> heading;
-	blaze::StaticVector<double, 3> target;
+	blaze::StaticVector<float, 3> position;
+	blaze::StaticVector<float, 3> heading;
+	blaze::StaticVector<float, 3> target;
 	unsigned int leader = 0;
 };
