+++ C++ Library/PathfindingLib/PathfindingLib/Pathfinding.cpp
@@ -5,17 +5,17 @@
 #include <unordered_set>
 #include <limits>
 
-double EuclideanDistance(blaze::StaticVector<double, 3> a, blaze::StaticVector<double, 3> b) {
-	blaze::StaticVector<double, 3> direction = b - a;
+float EuclideanDistance(blaze::StaticVector<float, 3> a, blaze::StaticVector<float, 3> b) {
+	blaze::StaticVector<float, 3> direction = b - a;
 	return blaze::length(direction);
 }
 
-double HeuristicCostEstimate(PGraphNode* a, PGraphNode* b) {
+float HeuristicCostEstimate(PGraphNode* a, PGraphNode* b) {
 	return EuclideanDistance(a->position, b->position);
 }
 
-std::vector<blaze::StaticVector<double, 3>>* reconstruct_path(std::unordered_map<blaze::StaticVector<int, 3>, PGraphNode*, PGraphNode::IndexHash>* came_from, PGraphNode* final_node) {
-	std::vector<blaze::StaticVector<double, 3>>* result = new std::vector<blaze::StaticVector<double, 3>>;
+std::vector<blaze::StaticVector<float, 3>>* reconstruct_path(std::unordered_map<blaze::StaticVector<int, 3>, PGraphNode*, PGraphNode::IndexHash>* came_from, PGraphNode* final_node) {
+	std::vector<blaze::StaticVector<float, 3>>* result = new std::vector<blaze::StaticVector<float, 3>>;
 	PGraphNode* current = final_node;
 	result->push_back(current->position);
 	while (came_from->find(current->index) != came_from->end()) {
@@ -26,7 +26,7 @@ std::vector<blaze::StaticVector<double, 3>>* reconstruct_path(std::unordered_map
 	return result;
 }
 
-std::vector<blaze::StaticVector<double, 3>>* Pathfinding::a_star(PGraph * graph, blaze::StaticVector<double, 3> from, blaze::StaticVector<double, 3> to)
+std::vector<blaze::StaticVector<float, 3>>* Pathfinding::a_star(PGraph * graph, blaze::StaticVector<float, 3> from, blaze::StaticVector<float, 3> to)
 {
 	PGraphNode* from_node = graph->NodeAt(from);
 	PGraphNode* to_node = graph->NodeAt(to);
@@ -55,7 +55,7 @@ std::vector<blaze::StaticVector<double, 3>>* Pathfinding::a_star(PGraph * graph,
 				continue;
 			}
 
-			float tentative_g_score = g_score.at(current->index) + static_cast<float>(EuclideanDistance(current->position, neighbor->position));
+			float tentative_g_score = g_score.at(current->index) + EuclideanDistance(current->position, neighbor->position);
 			float current_g_score = FLT_MAX;
 			if (g_score.find(neighbor->index) != g_score.end()) {
 				current_g_score = g_score.at(neighbor->index);
@@ -66,7 +66,7 @@ std::vector<blaze::StaticVector<double, 3>>* Pathfinding::a_star(PGraph * graph,
 
 			came_from[neighbor->index] = current;
 			g_score[neighbor->index] = tentative_g_score;
-			f_score[neighbor->index] = g_score.at(neighbor->index) + static_cast<float>(HeuristicCostEstimate(neighbor, to_node));
+			f_score[neighbor->index] = g_score.at(neighbor->index) + HeuristicCostEstimate(neighbor, to_node);
 
 			if (open_set.Contains(neighbor)) {
 				open_set.Update(neighbor);
@@ -77,5 +77,5 @@ std::vector<blaze::StaticVector<double, 3>>* Pathfinding::a_star(PGraph * graph,
 		}
 	}
 
-	return new std::vector<blaze::StaticVector<double, 3>>;
+	return new std::vector<blaze::StaticVector<float, 3>>;
 }
